ArtSignGenerator/requirements.txt
pyqt5
pillow
requests